Improve coloring enablement control
Enables molecule to detect more environment variables used to
enable or disable coloring
