+++ lib/molecule/logger.py
@@ -20,33 +20,10 @@
 """Logging Module."""
 
 import logging
-import os
-import sys
-from typing import Any
 
 from rich.logging import RichHandler
 
-
-# Based on Ansible implementation
-def to_bool(a: Any) -> bool:
-    """Return a bool for the arg."""
-    if a is None or isinstance(a, bool):
-        return bool(a)
-    if isinstance(a, str):
-        a = a.lower()
-    if a in ("yes", "on", "1", "true", 1):
-        return True
-    return False
-
-
-def should_do_markup() -> bool:
-    """Decide about use of ANSI colors."""
-    py_colors = os.environ.get("PY_COLORS", None)
-    if py_colors is not None:
-        return to_bool(py_colors)
-
-    return sys.stdout.isatty() and os.environ.get("TERM") != "dumb"
-
+from molecule.console import console
 
 SUCCESS = 100
 OUT = 101
@@ -108,7 +85,9 @@ def get_logger(name=None) -> logging.Logger:
     logger = logging.getLogger(name)  # type: logging.Logger
     logger.setLevel(logging.DEBUG)
 
-    handler = RichHandler(show_time=False, show_path=False, rich_tracebacks=True)
+    handler = RichHandler(
+        console=console, show_time=False, show_path=False, rich_tracebacks=True
+    )
     logger.addHandler(handler)
     logger.propagate = False
